requirements.txt
click==6.7
feedparser==5.2.1
Flask==1.1.1
Flask-Assets==0.12
Flask-FlatPages==0.6
Flask-Script==2.0.6
formic-py3==0.1.0
Frozen-Flask==0.15
future==0.16.0
glob2==0.6
itsdangerous==0.24
Jinja2==2.11.1
livereload==2.5.1
Markdown==2.6.9
MarkupSafe==1.0
PyYAML==5.3.1
six==1.11.0
tornado==4.5.2
webassets==0.12.1
Werkzeug==1.0.1
